TodosApp/app/requirements.txt
sqlalchemy>=1.4.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
psycopg2-binary>=2.9.0